+++ packages/wallet-sdk/src/CoinbaseWalletSDK.test.ts
@@ -2,28 +2,32 @@ import { CoinbaseWalletProvider } from './CoinbaseWalletProvider';
 import { CoinbaseWalletSDK } from './CoinbaseWalletSDK';
 import { ProviderInterface } from ':core/provider/interface';
 import { getFavicon } from ':core/type/util';
-import { getInjectedCbWalletMobileBrowserProvider } from ':util/provider';
+import { CBWindow } from ':util/provider';
+
+const window = globalThis as CBWindow;
 
 jest.mock(':core/type/util');
 jest.mock(':util/provider');
 jest.mock('./CoinbaseWalletProvider');
+class MockCipherProviderClass {
+  public isCoinbaseBrowser = true;
+}
+
+const mockCipherProvider = new MockCipherProviderClass() as unknown as ProviderInterface;
 
 describe('CoinbaseWalletSDK', () => {
   test('@makeWeb3Provider - return Coinbase Injected Provider', () => {
-    const injectedProvider = {} as unknown as ProviderInterface;
-    (getInjectedCbWalletMobileBrowserProvider as jest.Mock).mockReturnValue(injectedProvider);
-
+    window.ethereum = mockCipherProvider;
     const SDK = new CoinbaseWalletSDK({
       appName: 'Test',
       appLogoUrl: 'http://coinbase.com/wallet-logo.png',
     });
 
-    expect(SDK.makeWeb3Provider()).toBe(injectedProvider);
+    expect(SDK.makeWeb3Provider()).toBe(mockCipherProvider);
   });
 
   test('@makeWeb3Provider - return new CoinbaseWalletProvider', () => {
-    (getInjectedCbWalletMobileBrowserProvider as jest.Mock).mockReturnValue(undefined);
-
+    window.ethereum = undefined;
     const SDK = new CoinbaseWalletSDK({
       appName: 'Test',
       appLogoUrl: 'http://coinbase.com/wallet-logo.png',
@@ -45,7 +49,6 @@ describe('CoinbaseWalletSDK', () => {
 
   test('@makeWeb3Provider - default values for metadata', () => {
     (getFavicon as jest.Mock).mockReturnValue('https://dapp.xyz/pic.png');
-    (getInjectedCbWalletMobileBrowserProvider as jest.Mock).mockReturnValue(undefined);
 
     const SDK = new CoinbaseWalletSDK({
       appName: '',
